Modularize home page components
