+++ src/components/Home.js
@@ -1,8 +1,15 @@
 import Intro from "./Intro";
+import HomePageMap from "./HomePageMap";
+import {Wrapper} from "../GlobalStyles";
+import LocationFilter from "./LocationFilter";
 
 const Home = () => {
     return (
-        <Intro/>
+        <Wrapper>
+            <Intro/>
+            <LocationFilter/>
+            <HomePageMap/>
+        </Wrapper>
     );
 };
