+++ src/components/LocationFilter/LocationFilter.styles.js
@@ -1,0 +1,12 @@
+import tw, {styled} from "twin.macro";
+
+export const FilterContainer = styled.div`
+  ${tw`
+    shadow
+    bg-white
+    text-center
+    flex
+    justify-center
+    items-center
+  `}
+`;
